Rename deploymentconfigurations.* -> DeploymentConfigurations.*
Apply Google style naming scheme.
